Removing next-page-tester as it no longer works
As of Next.js 12, the `next-page-tester` library no longer works.  It shouldn't be suggested here (or it should come with a disclaimer that it doesn't work with v12).  There are no plans to add support for v12 so this isn't something anyone should be directed to and rely on in the future.